ICS_Bitbucket_repo/cloudrun_source/reflection_evaluation_pipeline/requirements.txt
Flask==3.0.0
# PyPDF2
# gcsfs
google-cloud-firestore==2.13.0
google-cloud-storage==2.11.0
sqlalchemy==2.0.21
python-dotenv==1.0.0
pymysql==1.1.1
pytz==2023.3.post1
langchain
langchain-community 
langchain-core
google-cloud-tasks
google-api-python-client
oauth2client
firebase-admin
google-cloud-secret-manager
google-auth
google-auth-oauthlib
google-cloud-aiplatform==1.40.0